% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shark_population_dataset.R
\docType{data}
\name{shark_data}
\alias{shark_data}
\title{Shark Population Dataset}
\format{
A data frame with 5 rows and 5 variables:
\describe{
  \item{initial_population}{Initial population of sharks.}
  \item{growth_rate}{Growth rate of the shark population.}
  \item{carrying_capacity}{Carrying capacity of the environment for sharks.}
  \item{time_steps}{Number of time steps to simulate.}
  \item{threat_factor}{Parameter representing a potential threat or environmental factor.}
}
}
\source{
Fake data generated for illustrative purposes.
}
\usage{
shark_data
}
\description{
This dataset contains simulated data for shark populations based on various parameters.
}
\keyword{datasets}
